pyproject.toml
[project]
name = "democracy-elo"
version = "0.1.0"
description = "A project to compare democracies."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]

[dependency-groups]
dev = [
    "uvicorn>=0.41.0",
]
